requirements.txt
pydantic>=2.0.0
openai>=1.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
